requirements.txt
# FaceTracker
opencv-python
mediapipe
numpy
scipy

# GazeTracker
dlib
numpy==1.26.4
opencv-python==4.10.0.82
setuptools==70.0.0

# MoonDream # Back to newer version, due to package conflicts with gazetracker
accelerate==0.32.1
huggingface-hub==0.24.0
Pillow==10.4.0
pyvips-binary==8.16.0
pyvips==2.2.3
torch==2.5.1
transformers==4.44.0
gradio==4.38.1
datasets==3.2.0
editdistance==0.8.1